- Introducing Automated testing Record macro. - Active Users view enhanced - Company embedding the Adress book party

+++ focWebServer/src/com/foc/vaadin/gui/components/tableAndTree/TreeLevelColorCellStyleGenerator.java
@@ -46,7 +46,7 @@ public class TreeLevelColorCellStyleGenerator implements Table.CellStyleGenerato
 		String style = null;
 
 		try{
-			int ref = ((Long)itemId).intValue();
+			long ref = ((Long)itemId).longValue();
 			FTree tree = (FTree) treeWrapper.getFocData();
 			FNode node = tree.findNode(ref);
 			style = getStyle(node);
